// src/saju/enums/twelve-spirits.enum.ts
/**
 * 12신살(十二神殺) - 길흉을 나타내는 신살들
 */
export enum TwelveSpirits {
  // 길신 (吉神)
  CHEONDEOK = '천덕',      // 天德 - 천덕귀인
  WOLDEOK = '월덕',        // 月德 - 월덕귀인
  CHEONEUL = '천을',       // 天乙 - 천을귀인
  TAEEUL = '태을',         // 太乙 - 태을귀인
  NOKSIN = '녹신',         // 祿神 - 녹신
  YANGINEUL = '양인을',    // 羊刃 - 양인
  
  // 흉신 (凶神)
  MANGSINAL = '망신살',    // 亡神殺 - 망신살
  GEOPSSAL = '겁살',       // 劫殺 - 겁살
  JAEHWASAL = '재화살',    // 災禍殺 - 재화살
  GUHAESAL = '구해살',     // 咎害殺 - 구해살
  YUKHASAL = '육해살',     // 六害殺 - 육해살
  HYEONGSAL = '형살'       // 刑殺 - 형살
}